distribution | begin refactoring to repository pattern
