+++ apps/distribution/src/common/utils/subscription-filter.utils.spec.ts
@@ -1,5 +1,5 @@
-import { SubscriptionFilter } from '../../resources/subscription/entities/subscription-filter.entity';
-import { Subscription } from '../../resources/subscription/entities/subscription.entity';
+import { SubscriptionFilter } from '../../resources/subscription/repository/entities/subscription-filter.entity';
+import { Subscription } from '../../resources/subscription/repository/entities/subscription.entity';
 import { SubscriptionQuery } from '../classes/subscription-query.class';
 import { FilterJoinOps, FilterOps } from '../types/filter.type';
 import {
