+++ apps/distribution/src/resources/distribution-event/distribution-event.controller.spec.ts
@@ -14,7 +14,7 @@ import { DistributionEventController } from './distribution-event.controller';
 import { DistributionEventService } from './distribution-event.service';
 import { CreateDistributionEventDto } from './dto/create-distribution-event.dto';
 import { UpdateDistributionEventDto } from './dto/update-distribution-event.dto';
-import { DistributionEvent } from './entities/distribution-event.entity';
+import { DistributionEvent } from './repository/entities/distribution-event.entity';
 
 describe('DistributionEventController', () => {
   let controller: DistributionEventController;
