+++ apps/distribution/src/resources/distribution-event/distribution-event.controller.ts
@@ -19,7 +19,7 @@ import { DefaultRuleException } from '../../common/errors/default-rule.exception
 import { DistributionEventService } from './distribution-event.service';
 import { CreateDistributionEventDto } from './dto/create-distribution-event.dto';
 import { UpdateDistributionEventDto } from './dto/update-distribution-event.dto';
-import { DistributionEvent } from './entities/distribution-event.entity';
+import { DistributionEvent } from './repository/entities/distribution-event.entity';
 
 @ApiTags('Distribution Event')
 @Controller('distribution-event')
