+++ apps/distribution/src/resources/distribution-event/distribution-event.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { DistributionEventController } from './distribution-event.controller';
 import { DistributionEventService } from './distribution-event.service';
-import { DistributionEvent } from './entities/distribution-event.entity';
+import { DistributionEvent } from './repository/entities/distribution-event.entity';
 
 @Module({
   imports: [SequelizeModule.forFeature([DistributionEvent])],
