+++ apps/distribution/src/resources/distribution-event/distribution-event.service.spec.ts
@@ -6,13 +6,13 @@ import {
   createMockRepository,
 } from '../../../test/helpers/database.helper';
 import { DefaultRuleException } from '../../common/errors/default-rule.exception';
-import { DistributionRule } from '../distribution-rule/entities/distribution-rule.entity';
-import { SubscriptionFilter } from '../subscription/entities/subscription-filter.entity';
-import { Subscription } from '../subscription/entities/subscription.entity';
+import { DistributionRule } from '../distribution-rule/repository/entities/distribution-rule.entity';
+import { SubscriptionFilter } from '../subscription/repository/entities/subscription-filter.entity';
+import { Subscription } from '../subscription/repository/entities/subscription.entity';
 import { DistributionEventService } from './distribution-event.service';
 import { CreateDistributionEventDto } from './dto/create-distribution-event.dto';
 import { UpdateDistributionEventDto } from './dto/update-distribution-event.dto';
-import { DistributionEvent } from './entities/distribution-event.entity';
+import { DistributionEvent } from './repository/entities/distribution-event.entity';
 
 describe('DistributionEventService', () => {
   let service: DistributionEventService;
