+++ apps/distribution/src/resources/distribution-event/repository/entities/distribution-event.entity.ts
@@ -1,6 +1,6 @@
 import { Column, DataType, HasMany, Model, Table } from 'sequelize-typescript';
-import { DistributionRule } from '../../distribution-rule/entities/distribution-rule.entity';
-import { Subscription } from '../../subscription/entities/subscription.entity';
+import { DistributionRule } from '../../../distribution-rule/repository/entities/distribution-rule.entity';
+import { Subscription } from '../../../subscription/repository/entities/subscription.entity';
 
 @Table({ tableName: 'distribution_event' })
 export class DistributionEvent extends Model {
