+++ apps/distribution/src/resources/distribution-log/distribution-log.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { DistributionLogController } from './distribution-log.controller';
 import { DistributionLogService } from './distribution-log.service';
-import { DistributionAttempt } from './entities/distribution-attempt.entity';
-import { DistributionLog } from './entities/distribution-log.entity';
+import { DistributionAttempt } from './repository/entities/distribution-attempt.entity';
+import { DistributionLog } from './repository/entities/distribution-log.entity';
 
 @Module({
   imports: [SequelizeModule.forFeature([DistributionLog, DistributionAttempt])],
