+++ apps/distribution/src/resources/distribution-log/repository/entities/distribution-attempt.entity.ts
