+++ apps/distribution/src/resources/distribution-rule/distribution-rule.module.ts
@@ -3,7 +3,7 @@ import { SequelizeModule } from '@nestjs/sequelize';
 import { DistributionEventModule } from '../distribution-event/distribution-event.module';
 import { DistributionRuleController } from './distribution-rule.controller';
 import { DistributionRuleService } from './distribution-rule.service';
-import { DistributionRule } from './entities/distribution-rule.entity';
+import { DistributionRule } from './repository/entities/distribution-rule.entity';
 
 @Module({
   imports: [
