+++ apps/distribution/src/resources/subscription/repository/entities/subscription.entity.ts
@@ -7,12 +7,12 @@ import {
   Model,
   Table,
 } from 'sequelize-typescript';
-import { FilterJoinOps } from '../../../common/types/filter.type';
+import { FilterJoinOps } from '../../../../common/types/filter.type';
 import {
   SubscriptionData,
   SubscriptionType,
-} from '../../../common/types/subscription-type.type';
-import { DistributionEvent } from '../../distribution-event/entities/distribution-event.entity';
+} from '../../../../common/types/subscription-type.type';
+import { DistributionEvent } from '../../../distribution-event/repository/entities/distribution-event.entity';
 import { SubscriptionFilter } from './subscription-filter.entity';
 
 @Table({
