+++ apps/notification/src/resources/email-template/email-template.controller.ts
@@ -13,10 +13,10 @@ import {
 } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import * as _ from 'lodash';
-import { EmailTemplate } from './repository/entities/email-template.entity';
 import { CreateEmailTemplateDto } from './dto/create-email-template.dto';
 import { UpdateEmailTemplateDto } from './dto/update-email-template.dto';
 import { EmailTemplateService } from './email-template.service';
+import { EmailTemplate } from './repository/entities/email-template.entity';
 
 @ApiTags('Email Template')
 @Controller('email-template')
